day07: The foundations: a flat table of paths.
I started to make a pointery tree structure (which, I *can*, I mean I
could do that in C if I had to), and then I decided that was too much.

There might be a simpler way to do this just by scanning through the
text but at this point I don't care.