day09: Yes of course there was a bug.
Pride goeth, etc.  I more or less knew what I was trying to do with
`contract()`, but if I'd tried explaining anything whatsoever about this
in prose I might've caught that the return value was backwards; I want
to know how the tail moves when acting as the rest of the rope's head
and not, like, what to insert behind it to keep the rest of the rope
static.

(Also I'm realizing I could've made `contract()` functional and gotten
the next `Move` by subtracting the new and old values — which is kind
of screamingly obvious when I write it down like that — but here we
are.)